refactor(mcp-server): split tool decision tests
